// src/app/modules/budget-tracker/modules/dashboard/modules/info-cards/models/info-card-value-modal-data.ts
import { InfoCardValueToEdit } from './info-card-value-to-edit.enum';
import { InfoCardMenuActionsType } from './info-card-menu-actions-type.enum';

export interface InfoCardValueModalData {
  actionType: InfoCardMenuActionsType;
  valueToEdit: InfoCardValueToEdit;
  initialValue: number;
  shouldDisplayInitialValue?: boolean;
}